requirements.txt
Flask==2.0.3
werkzeug==2.0.3
requests
opencv-python
pyzbar
python-dotenv
google-genai
